Add NAICS 2 digits table
