+++ src/StephJob/Core/Repositories/INaics2DigitRepository.cs
@@ -1,0 +1,8 @@
+﻿namespace StephJob.Core.Repositories
+{
+    using StephJob.Core.Domain;
+
+    public interface INaics2DigitRepository : IRepository<Naics2Digit>
+    {
+    }
+}
