+++ src/StephJob/Persistance/StephContext.cs
@@ -21,6 +21,8 @@
 
         public DbSet<Naics> Naics { get; set; }
 
+        public DbSet<Naics2Digit> Naics2Digits { get; set; }
+
         public DbSet<Noc2011> Noc2011s { get; set; }
 
         public DbSet<Sdgs372> Sdgs372s { get; set; }
