+++ src/StephJob/Persistance/UnitOfWork.cs
@@ -15,6 +15,8 @@
 
         private readonly Lazy<INaicsRepository> naics;
 
+        private readonly Lazy<INaics2DigitRepository> naics2Digits;
+
         private readonly Lazy<INoc2011Repository> noc2011s;
 
         private readonly Lazy<ISdgs372Repository> sdgs372s;
@@ -40,6 +42,7 @@
             this.employmentAveWeekEarnings = new Lazy<IEmploymentAveWeekEarningRepository>(() => new EmploymentAveWeekEarningRepository(context));
             this.futureEmployments = new Lazy<IFutureEmploymentRepository>(() => new FutureEmploymentRepository(context));
             this.naics = new Lazy<INaicsRepository>(() => new NaicsRepository(context));
+            this.naics2Digits = new Lazy<INaics2DigitRepository>(() => new Naics2DigitRepository(context));
             this.noc2011s = new Lazy<INoc2011Repository>(() => new Noc2011Repository(context));
             this.sdgs372s = new Lazy<ISdgs372Repository>(() => new Sdgs372Repository(context));
             this.sdgs411s = new Lazy<ISdgs411Repository>(() => new Sdgs411Repository(context));
@@ -52,11 +55,12 @@
         }
 
         public IEmploymentAveWeekEarningRepository EmploymentAveWeekEarnings => employmentAveWeekEarnings.Value;
-
-
+        
         public IFutureEmploymentRepository FutureEmployments => futureEmployments.Value;
 
         public INaicsRepository Naics => naics.Value;
+
+        public INaics2DigitRepository Naics2Digits => naics2Digits.Value;
 
         public INoc2011Repository Noc2011s => noc2011s.Value;
